pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cwi-accountant"
version = "0.1.0"
description = "Local-first accounting agent and human review dashboard for CWI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.8,<3",
  "typer>=0.12,<1",
  "watchdog>=4,<5",
  "openpyxl>=3.1,<4",
  "python-dateutil>=2.9,<3",
  "rapidfuzz>=3.9,<4",
  "streamlit>=1.37,<2",
  "pandas>=2.2,<3",
  "PyYAML>=6,<7",
  "pypdf>=4,<6",
  "pillow>=10,<12",
]

[project.optional-dependencies]
ocr = ["pytesseract>=0.3.10,<0.4"]
dev = [
  "pytest>=8.2,<9",
  "pytest-cov>=5,<6",
  "mypy>=1.10,<2",
  "ruff>=0.5,<1",
]

[project.scripts]
cwi-accountant = "cwi_accountant.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "S"]
ignore = ["S101"]
